api/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
python-dotenv==1.0.1
redis==7.0.1
pytest==9.0.1
pytest-asyncio==1.3.0
sqlalchemy==2.0.44
alembic==1.17.2
psycopg2-binary==2.9.11
passlib == 1.7.4